wip. Comments and insert logic modified
Insert now can traverse forwards and backwards for better performance, depending on where the index is placed. Backwards or tail traverse makes an extra step to have the tail pointer placed before the node to be replaced, and this be pushed forward. This is intended to keep similar logic on the insertion logic for both forward and backward traverse. Tests are pending