WIP. span added dynamically to scroll-content console.
There are still very few things to polish. Check for D&Drop API to include insert and prepend logic to DOM elements, check for insertBefore and nextSibling, use pointers and event triggerer could be dragover, but targeting the img, not the container. How does the dragged element knows it is over another img that has o does not have siblings. source code for tiermaker?