Refactor + changes. Insert logic now works with remove logic.
Finally, the work is done. There are still same edge-cases for when the nodes are dropped in the same tier in edges
or places where no drop zone exists. Yet I think a no-compromise solution is still capable (otherwise, prepare disclaimer).

Basically all addition and substraction of nodes is done in the drop handler. This allows for much better control
over the data structures and is more immediate for detectig changes, and avoids the previous logic to be handled
in the dragend event handler (which surprisingly is called AFTER the drop event. Note to look deeper into this).

All hard work seems to be done now, small modifications, utils for DRY and better stylings are pending.